feat(runtime): C5 AX-core delta sync — swap_check emitter, CertStore, first-fault envelope
Propagates the monorepo course-correction Phase C delta (archive aad96b7a..fff42708):
W-A FirstFaultEnvelope + swap_check emitter at the mode.rs seam (G-3 fix), capped
mode-gated CertStore, sizing tests; lockstep re-pin to the C5 train revs. Also adds
the docs/spec/schemas slice that the cert schema-pinned test reads via a
CARGO_MANIFEST_DIR-relative path — a Phase-B verification gap surfaced by this
wave and fixed forward here (the test could not have passed without the slice).
Gates green standalone: fmt --check, clippy --all-targets -D warnings, cargo test.

+++ crates/mycelium-cert/tests/mutant_witnesses.rs
@@ -31,7 +31,7 @@ fn byte_of(value: i64) -> Value {
     .unwrap()
 }
 
-fn tern_of(value: i64, trits: u32) -> Value {
+fn tern_of(value: i128, trits: u32) -> Value {
     Value::new(
         Repr::Ternary { trits },
         Payload::Trits(ternary::int_to_trits(value, trits).unwrap()),
